rename utils.py -> FSUtils.py
the old name had a good risk of causing collions with some other module
importing it's own utils file.
